first pass at meta
