+++ package.json
@@ -11,6 +11,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-error-boundary": "^3.1.4",
+    "react-helmet": "^6.1.0",
     "react-icons": "^4.4.0",
     "react-popper": "^2.3.0",
     "react-query": "^3.39.1",
@@ -40,6 +41,7 @@
     "@types/node": "16.11.43",
     "@types/react": "18.0.14",
     "@types/react-dom": "18.0.5",
+    "@types/react-helmet": "^6.1.5",
     "change-case": "4.1.2",
     "cross-env": "7.0.3",
     "dependency-cruiser": "11.11.0",
