+++ public/index.html
@@ -5,6 +5,11 @@
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
+    <meta property="og:site_name" content="Justin Klaas Portfolio"/>
+    <meta property="og:title" content="Justin Klaas Portfolio"/>
+    <meta property="og:type" content="website"/>
+    <meta name="description" content="Justin Klaas is an application developer specializing in React-Native, GraphQL, and Typescript." />
+    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
