+++ src/components/Providers/index.tsx
@@ -1,0 +1,10 @@
+import { ReactQueryClientProvider } from "./ReactQueryClientProvider"
+import { RouteProvider } from "./RouteProvider"
+
+export function Providers() {
+  return (
+    <ReactQueryClientProvider>
+      <RouteProvider />
+    </ReactQueryClientProvider>
+  )
+}
